+++ CHANGELOG.md
@@ -9,7 +9,14 @@
 
 NOTE:
 
+* [BUGFIX] marshal.go: improve packet validation and error handling #323
+* [BUGFIX] marshal.go: Fix on-error-continue flow in sendOneRequest #324
+
+## v1.31.0
+
 * [BUGFIX] Add validation to prevent calling updatePktSecurityParameters with non v3 packet #251 #314
+* [ENHANCEMENT] walk.go: improve BulkWalk error handling #306
+* [ENHANCEMENT] return received SNMP error code immediately instead of waiting for timeout #319
 
 ## v1.30.0